Fixed while and optimised compiler (by passing bc as a reference).
Fixed bug in jump behaviour, where the destinations of the jumps weren't
calculated correctly, by patchingthe correct destination in later. Also
introduced the `Patch` opcode to make it more obvious when patching
jumps goes wrong in the compiler. While we are here, also simplified the
code by passing `bc` in as a reference.